# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: AegisCode DevSecOps API
  description: Shift-Left Multi-Agent Architecture API para análisis estático, ingesta de repositorios Git y mitigación automatizada de vulnerabilidades.
  version: 1.0.0
  contact:
    name: AegisCode Team
servers:
  - url: http://localhost:8080
    description: Entorno de Desarrollo Local
paths:
  /api/auth/login:
    post:
      operationId: authLogin
      summary: Inicia sesion de usuario
      description: Autentica un usuario con username y password, devuelve un token JWT.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Autenticacion exitosa.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Credenciales invalidas.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/register:
    post:
      operationId: authRegister
      summary: Registra un nuevo usuario
      description: Crea un nuevo usuario en la base de datos con password hasheada.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registro exitoso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '409':
          description: El usuario ya existe.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido via /api/auth/login
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 100
          example: "admin"
        password:
          type: string
          minLength: 1
          maxLength: 255
          example: "securepassword123"
    RegisterRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 100
          example: "newuser"
        password:
          type: string
          minLength: 6
          maxLength: 255
          example: "mypassword"
        scopes:
          type: array
          description: Scopes del usuario. Opcional, por defecto ["READER"].
          items:
            type: string
            enum: [READER, WRITER]
          default: ["READER"]
          example: ["READER", "WRITER"]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: Token JWT de acceso.
          example: "eyJhbGciOiJIUzI1NiJ9..."
        username:
          type: string
          example: "admin"
        scopes:
          type: string
          description: Scopes del usuario separados por coma (READER, WRITER, etc.)
          example: "READER,WRITER"
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        username:
          type: string
          example: "admin"
        scopes:
          type: string
          description: Scopes del usuario separados por coma.
          example: "READER,WRITER"
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Mensaje de error descriptivo.
          example: "Credenciales invalidas"
...